+++ src/models/Hotelreservierung.java
@@ -1,6 +1,7 @@
 package models;
 
 import utils.Input;
+import utils.Utils;
 
 import java.time.Duration;
 import java.util.Date;
@@ -13,12 +14,11 @@ public class Hotelreservierung extends Reservierung {
     String hotelname;
     Duration reisedauer;
 
-    public Hotelreservierung(Date date, int summe, String hotelname, Duration reisedauer) {
+    public Hotelreservierung(Date date, double summe, String hotelname, Duration reisedauer) {
         super(date, summe);
         this.hotelname = hotelname;
         this.reisedauer = reisedauer;
     }
-
 
     public static Hotelreservierung readHotelreservierungFromInput(Input input) {
         Date datum = input.readDate("Reservierungsdatum: ");
@@ -27,7 +27,6 @@ public class Hotelreservierung extends Reservierung {
         int duration = input.readInt("Reisedauer (in Tagen): ", 1, Integer.MAX_VALUE);
         return new Hotelreservierung(datum, summe, hotelname, Duration.of(duration, DAYS));
     }
-
 
     public String getHotelname() {
         return hotelname;
